embed IPython and add extra info to maze tiles.
The IPython is there to help view more data about the maze.
During execution, I can see the info levels of the tiles, potentially
the graph that the robot is building, etc. Also, I added in a knowledge
index. For each tile, there are 4 different directions. The robot
should strive to discover the maze, and a knowledge index helps it
determine which areas are undiscovered. The KI is 0-4, indicating
which sides are known.