fixed bug regarding the view for maze tile connections.
The maze printer only checked to see if the key was in the dictionary.
It didn't check to see if the result was None, so it would print a
connection, despite one not being there. This was an interpretation
fault, I should have hammered out into the code that the key might
exist, but that doesn't mean a connection exists.