Gain more data on sensor input.
Now, instead of knowing only which tiles are conected,
we know which tiles are not connected. If we see 3 forward,
we know there is a wall at the end, or the end of the maze.
Our robot now updates the tile on the other side, citing to know
something about it, or ignores it, since it's the edge of the maze.