Determine waypoints to explore the maze.
A lot of stuff happened in this commit. Many little debug
functions were introduced so that data could be more easily
seen during the IPython embed. The robot now properly creates
the graph of the maze and tile transitions. The robot also
calculates waypoints to figure out the next best node to
investigate.