robot can determine path of squares to the waypoint.
The robot does a breadth first search to find the
waypoint. As soon as it sees it, it builds a list
of tiles that form a path. With this list, it is
easy to determine the robots next move. It simply
out how they are linked, and does the appropriate
action to get to the next tile on the path. We will
implement this next.