Add performance metrics and improve algo
The robot now records how much of the maze it knows,
and updates the value to a file so that data can be
later processed. This is the performance metrics folder.
The robot also has an algo change - it decides between two
candidate tiles to explore, deciding to go with a tile that
has greater distance from the center if it is sufficiently
closer to the robot.