Robot can correctly use full range of motion
The robot is allowed to make a rotation, and then
move up to 3 steps. Before, the robot would only
ever take one step. But now, it can figure out if it
can skip some steps. This greatly helps improve the
time the robot spends searching the maze. (Around 5
points for 12 x 12 and 7 points for 16 x 16) Assuming
the robot explores the maze to completion.