Robot now prefers immediate searches over long range ones
With the realization that a tile furthest away from the
center can still be monumental to finding the shortest
path, I changed the algo so that it favors searching
tiles in its immediate proximity over tiles that are
far away but closer to the center. The results are our
fastest times yet, and it is reasonable to expect the
robot to solve the maze.