Cut out redundant computation
The robot's front mounted sensor is almost completely useless
The only time it is ever used is in the beginning to see how
far it can move forward. Every other time, the robot will know
how much it sees in the front because it would have had to turn
into that view, which means it would have seen the distance
with its side sensors.