Complete the maze update logic with full move flexibility
The robot will now update the maze perfectly, giving each
tile a dict to every possible location it could go to in
one action. It does this efficiently too, considering just
the tiles that will likely need to be updated and don't
have the information already. This is a max of 12 tiles in
any given case.