Fix dijkstra's algorithm and create a manual test run
As of right now, we're testing the robot's ability to
perfectly use the data it has collected in the first run.
There should be no real algorithms or anything else
involved here, except for running dijkstra's correctly.
Hence, a manual run through the maze was created to
test this feature.