Implement BFS and dijkstra graph walking exploration algos.
After using copypasta to use dijkstra algo to explore
the maze, the robot did make it to the end of the run,
but the computation itself took way too much time.
Simplifying it using BFS was a little better. This gave
us two different algos we can use to solve the maze. Both
do almost the same thing, it's just that in certain situations,
due to the way keys are ordered in dicts, the two algos
will choose differnet destination paths that are the
same distance away. This is what leads to the slight
variation in scores.