Fix the djikstra algorithm bug.
For some reason, the code didn't add the destination to the
set of visited nodes when building the set of connected
nodes. The only time the node would be added is when the node
itself was the current node. This behavior caused long run times
and in one particular graph, caused an infinite loop. This loop
was what alerted me to the possibility that this is a bug, and
the algorithm isn't just one that takes a long time.