implement the legacy solutions that we had.
This was the center hunter and the single step center hunter.
The center hunter would prioritize tiles near the center,
above all else. The single step center hunter would do the
same, but was limited in that it could only ever take one
step at a time. These were implemented so that we could
see how our refinements have improved the algorithm runtime.